requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.5
transformers
torch
yake
keybert
